backend/tile_3/edubot_blob_cosmos/requirements.txt
azure-functions
transformers
language-tool-python
azure-storage-blob
azure-cosmos
torch
language-tool-python
scikit-learn
spacy
rake-nltk
sentence_transformers
nltk
python-dotenv